+++ src/main/java/com/peasenet/mods/combat/ModAutoAttack.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2022. Gavin Pease and contributors.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+ *  of the Software, and to permit persons to whom the Software is furnished to do so, subject to the
+ *  following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+ * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT
+ * OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+
+package com.peasenet.mods.combat;
+
+import com.peasenet.mods.Mod;
+import com.peasenet.mods.Type;
+import com.peasenet.util.PlayerUtils;
+import net.minecraft.util.hit.EntityHitResult;
+
+/**
+ * A mod that makes the player attack the entity that it is currently looking at.
+ */
+public class ModAutoAttack extends Mod {
+
+    public ModAutoAttack() {
+        super(Type.AUTO_ATTACK);
+    }
+
+    @Override
+    public void onTick() {
+        var target = getClient().crosshairTarget();
+        if (!(target instanceof EntityHitResult))
+            return;
+        var entity = ((EntityHitResult) target).getEntity();
+        PlayerUtils.attackEntity(entity);
+    }
+
+}
