+++ src/main/java/com/peasenet/mods/movement/ModDolphin.java
@@ -4,9 +4,8 @@
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction, including
  * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- *  of the Software, and to permit persons to whom the Software is furnished to do so, subject to the
- *  following conditions:
- *
+ * of the Software, and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
  * The above copyright notice and this permission notice shall be included in all copies or substantial
  * portions of the Software.
  *
@@ -18,30 +17,27 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-
-package com.peasenet.mods.combat;
+package com.peasenet.mods.movement;
 
 import com.peasenet.mods.Mod;
 import com.peasenet.mods.Type;
-import com.peasenet.util.PlayerUtils;
-import net.minecraft.util.hit.EntityHitResult;
+import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
 
-/**
- * A mod that makes the player attack the entity that it is currently looking at.
- */
-public class AutoAttack extends Mod {
-
-    public AutoAttack() {
-        super(Type.AUTO_ATTACK);
+public class ModDolphin extends Mod {
+    public ModDolphin() {
+        super(Type.DOLPHIN);
     }
 
     @Override
     public void onTick() {
-        var target = getClient().crosshairTarget();
-        if (!(target instanceof EntityHitResult))
+        super.onTick();
+        if (getPlayer() == null)
             return;
-        var entity = ((EntityHitResult) target).getEntity();
-        PlayerUtils.attackEntity(entity);
+        // check if the player is swimming
+        if (getPlayer().isTouchingWater()) {
+            // if they are, move them up
+            getPlayer().addVelocity(0, 0.1, 0);
+            getPlayer().networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(getPlayer().getX(), getPlayer().getY() + 0.1, getPlayer().getZ(), false));
+        }
     }
-
 }
